Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Here's the assignee usage/user for the past 30 days on our Gerrit instance with around 700-800 active users.
> Around 10% uses it each month, which is as expected since it's only used when needed.
> The set of users vary for each individual 30 day period as well as their usage frequency.
> Over a 6 month period more than 100 users have used the assignee feature.
> 
> These are the users that used it and how often they've used it the past 30 days:
> 
> https://imgur.com/a/JaTfByX
> 
> We have not advertised the feature other than a notice that the feature is available when we upgraded to a version that supported it. It was adopted from the beginning has grown organically.

Thanks for the info. That is roughly aligned with what we see (about 5% of changes having an assignee at least once). The real question, though, is: does it actually help with the problem that users have or is it only slightly improving on a situation that could otherwise be better. I guess what I am trying to say is: our data suggest that the root cause of the problem we are trying to solve with the assignee (and the attention set) is actually the belief that the review will go faster by inviting more reviewers. 

The attention set intends to make reviews go faster by making the dashboard more useful (and, indirectly, email notification too). One core assumption of the attention set is that it will work very well when we get users to not over-invite (as much), as everything in the attention set is supposed to be relevant (just like an assigned change today). That can only work when people refrain from over-inviting. In that case, we believe (and know from our internal tool) it will actually work better than assignee, as it will allow change owners to provide useful priority signals to multiple users at the same time.

Another part of the solution then also is to advocate against this behavior among our users. Additionally, we will also need to add features that will help them to solve the need for speedy reviews by selecting a narrow set of "good" reviewers to begin with (better owner support, availability indicators, etc.).

Patch-set: 10