Update patch set 10
Patch Set 10:

Thanks for the clarification. I think we are mostly on the same page then. Keeping the assignee feature alive should not be a big problem.

I am still in doubt whether it has enough value when used in combination with the attention set, but I am fine with every host admin to decide that for themselves.

I think you are using assignee then mostly for expressing urgency and priority, which is a separate (valid!) use case. Two ideas that I have in mind for this:

1. Add a "priority" feature such that owners can express how important/urgent something is.
2. Automatically sorting the dashboard by how long you have already been in the attention set.

Your example of 20 changes with 12 of them requiring your attention is very good to understand where you are coming from. Thanks for that!

Patch-set: 10