Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > Here's the assignee usage/user for the past 30 days on our Gerrit instance with around 700-800 active users.
> > Around 10% uses it each month, which is as expected since it's only used when needed.
> > The set of users vary for each individual 30 day period as well as their usage frequency.
> > Over a 6 month period more than 100 users have used the assignee feature.
> > 
> > These are the users that used it and how often they've used it the past 30 days:
> > 
> > https://imgur.com/a/JaTfByX
> > 
> > We have not advertised the feature other than a notice that the feature is available when we upgraded to a version that supported it. It was adopted from the beginning has grown organically.
> 
> Thanks for the info. That is roughly aligned with what we see (about 5% of changes having an assignee at least once). The real question, though, is: does it actually help with the problem that users have or is it only slightly improving on a situation that could otherwise be better. I guess what I am trying to say is: our data suggest that the root cause of the problem we are trying to solve with the assignee (and the attention set) is actually the belief that the review will go faster by inviting more reviewers. 
> 
> The attention set intends to make reviews go faster by making the dashboard more useful (and, indirectly, email notification too). One core assumption of the attention set is that it will work very well when we get users to not over-invite (as much), as everything in the attention set is supposed to be relevant (just like an assigned change today). That can only work when people refrain from over-inviting. In that case, we believe (and know from our internal tool)

I did not know that your internal tool used assignee before it used attention-set. If that's the case it would be interesting to get some data on review-time before and after you switched from assignee to attention-set.
If that is not the case you cannot possibly know that it will work better than assignee.

Attention-Set should be compared with the current "reviewed-flag" in Gerrit which tries to give the information "it's time to look at this change again".
Compared to the current "reviewed-flag" attention-set will certainly decrease review-time as the "reviewed-flag" is inferior to attention-set when it comes to trying to determine whether the state of a change has changed in such a way that it needs to look at it again.

> it will actually work better than assignee, as it will allow change owners to provide useful priority signals to multiple users at the same time.

I'm starting to wear down my fingers trying to explain that attention-set and assignee are not the same thing, and are different, non-conflicting, solutions to a similar or same problem.

I am NOT doubting that Attention-Set will decrease review-time on average.
What I am saying is that removing assignee will *increase* review-time, with or without attention-set. And it will first and foremost impact the changes with the longest review-times and the changes that are time and/or resource critical the most, relative to review-time for the same changes *with* assignee.

Review-Time:
______________________________________
|               | Nothing | Assignee |
--------------------------------------
| Reviewed-flag |  HIGH   | LOWER    |
--------------------------------------
| Attention-Set | LOWER   | LOWEST   |
--------------------------------------



What I have tried to describe repeatedly is that assignee and AS has different levels of priority signals:
AS gives higher priority to f.i. a change were the reviewer hasn't reviewed the latest patchset compared to changes the reviewer has already reviewed the latest patchset.
However!
AS gives all the individual reviewers in the AS of a particular change the same priority.
AS gives the same priority, from the reviewers perspective, for each change where that specific reviewer is a member of the AS.

You should review all changes where you are added to the AS, but you should start with the changes for which you are assignee.

This extra priority-level (assignee) is helpful even outside of the code-review context.
You might be added to the attention-set of 4 changes, but you have other tasks that are prioritized above general code-review ATM.
If you are assigned to one of those code-review you might do this code-review before the other task.

1. Do critical code-review
2. Put out fire in burning production system
3. Do code-review for the 3 remaining changes.

> 
> Another part of the solution then also is to advocate against this behavior among our users. Additionally, we will also need to add features that will help them to solve the need for speedy reviews by selecting a narrow set of "good" reviewers to begin with (better owner support, availability indicators, etc.).

Yes, yes and yes.
This does not negate the benefits of assignee.

Patch-set: 10
Reviewer: Gerrit User 1014252 <1014252@173816e5-2b9a-37c3-8a2e-48639d4f1153>