Update patch set 10
Patch Set 10:

Hi Sven, just wanted to get back to you on this particular part.

> 
> Let me begin by clarifying that I have never stated, or suggested, that adding more reviewers leads to faster review-times, I have repeatedly made the reverse statement.
> 
> > Patch Set 10:
> > 
> > > UC1
> > > I need reviews from (A|B|C)&D
> > > I would "request action" from A,B and C. Since it's, on average, three times as likely to get a timely review from three people than from one.
> > > I would "demand action" from D.

I have a hard time reconciling your first sentence above with what you wrote in UC1. To me, this reads like if you only invited A&D you believe that on average there is a 2 in 3 chance that this review will go slower. Which to me is equivalent to saying that adding more people makes it more likely to go quicker. Are you just saying that is what you see users do and not what you believe is a valid use case?

> > 
> > Sven, fyi, we did some research on our Gerrit logs and the results suggest that the behavior you describe here is actually associated with reviews that take longer (+5 hours time to submit for the median change having vs. not having at least one reviewer invited that did not end up voting).
> 
> These numbers are describing the fact that if a change doesn't get a timely review the owner is likely to invite additional reviewers.
> 
> A simple and, IMO, not uncommon use-case:
> You upload a change and add reviewer A.
> Reviewer A is otherwise engaged and after 5 hours without activity you add reviewer B.
>  

I just re-ran some of the data to check on that. Thanks for the hint. The effect actually stays the same when you restrict the set of invited reviewers to the timestamp of the very first invite. So I don't think the above is a valid explanation.

In any case, I appreciate your patience with explaining your point of view and am looking forward to talking more in person tomorrow.

Patch-set: 10