Update patch set 10
Patch Set 10:

(6 comments)

ben, your frustration is noted, and be assured that it goes both ways. ;)

please be careful to not throw around "upload your own design doc" too lightly - you did this here with both me and sven (and someone else did it with me in another design review), and all you're going to get from that is disengagement - people don't necessarily have the time to come up with a fully-fledged proposal of their own, and they are even less inclined to do so if they get the impression that they are not being taken seriously to start with. that makes it effectively a cudgel to shut down criticism.

consequently, i'd ask you to do the following:
- where valid points have been raised, make sure to work them into the design document properly - as alternatives, as cons, as inline discussions, whatever seems right. _in particular_ when they pertain to fundamental aspects of the proposal. please go again through all previous comments with this in mind.
- make sure that you _honestly_ consider alternatives, that is, you try to find ways to address the issues you have identified in the existing features without discarding them altogether, especially where it was already pointed out to you that the issue is merely an implementation detail (missing visualization, missing "knob" in a relevant place, etc.). otherwise you're effectively straw-manning, and that doesn't help. only when you can conclusively show that relevant issues cannot be addressed incrementally, but they can with your proposal, you have made a case for adding complexity and/or putting everything upside down.

anyway, my fundamental contention with your approach is that it takes power out of my hands (in particular when notification mails are suppressed; otherwise it's just totally useless). *i* want to decide what i want to follow up to - it's not for others or a machine to decide that for me. of course, i also want to be efficient, so i want pertinent indicators that enable me to make a "no" or "later" decision "at a glance" in _most_ cases.
i have already presented my ideas how to achieve that; please ask _concrete_ questions about them - i'm sure that in most cases you'll find the answer yourself.

regarding the assignee feature, the way sven describes it (i have no experience myself; qt's gerrit was way behind until recently), it is used basically as a band-aid when the process has already broken down, and somewhat unsurprisingly the google numbers suggest that it doesn't actually help.
my gut feeling is that a more elaborate approach like the one proposed here won't do better relative to the baseline of an actually functional "reviewed" feature as i'm proposing.

regarding the proposed telco, i'm theoretically interested, but i'm not very good at real-time comms that require actual thinking, and telcos in particular tend to be a pita. also, the proposed date effectively clashes with https://fridaysforfuture.de/neustartklima/

Patch-set: 10