Update patch set 10
Patch Set 10:

(5 comments)

> Please be careful to not throw around "upload your own design doc" too lightly.

Thanks for the advice. I have done that, because

- It is part of the process that the community has defined.
- I had a hard time understanding what you actually prefer over what I am proposing.
- It is great for other reviewers to have a concise alternative proposal.
- I am not in a position to write down your alternative, because I don't understand it and don't believe in it.
- Compared to the time that it must have taken you to review this change so thoroughly I would guess writing an alternative proposal should take much less time.

> Where valid points have been raised, make sure to work them into the design document properly - as alternatives, as cons, as inline discussions ...

I don't think "inline discussions" are a good idea. The proposal should be concise on its own.

"Alternatives" should be written by someone else. I find it really hard to write an alternative proposal that I don't fully understand and don't believe in.

"Cons" is similar to "Alternatives". Basically I have been trying to convince my reviewers that all their concerns are no "Cons" at all. But feel free to suggest a paragraph that I would be happy to add to "Cons" and attribute it to you.

> Please go again through all previous comments.

Sorry, I have already gone through all comments and replied to them. That is all I can do.

> Make sure that you _honestly_ consider alternatives.

I have done that. Sorry that this did not come across.

> You have made a case for adding complexity and/or putting everything upside down.

You are welcome to vote -1 and I am sure that the ESC will consider your concerns appropriately. I was not aiming for universal consensus. It is fine to disagree on a matter.
 
> Regarding the proposed telco, i'm theoretically interested, but i'm not very good at real-time comms that require actual thinking, and telcos in particular tend to be a pita. also, the proposed date effectively clashes with https://fridaysforfuture.de/neustartklima/

Feel free to propose another date. I am happy to meet next week, as well.

I understand that this review may look as if Sven and you are my only serious reviewers of an idea that I have come up with over a weekend. However, that is a false impression. The proposal was presented at the Gerrit User Summit and discussed with representatives from Android, Chromium, CollabNet, GerritForge, Qualcomm, Google's Gerrit team and others.

And note that "Google’s internal code review system Critique launched an "Attention Set" feature in 2015 [...]. It was extremely well received and has become a vital part of the code review system." (see use-cases.md).

So what I am proposing is I think very well tested and reviewed already. Maybe that is why I come across a bit overconfident. Sorry for that.

Patch-set: 10