Update patch set 12
Patch Set 12:

you may interpret the "you" as referring to google, not ben personally. doesn't change my perception of things. take it as a reminder that as an entity with a disproportionate influence google needs to work extra hard to be transparent if the open governance model is to be credible. (i have no good things to say in that regard about my former employer, the qt company - it's one of the reasons why i left.)

Patch-set: 12
Reviewer: Gerrit User 1013267 <1013267@173816e5-2b9a-37c3-8a2e-48639d4f1153>