Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I discussed this with Eryk.
> 
> - We're not opposed to moving this functionality to a plugin.  This would actually make it more consistent with project deletion which is already a (core) plugin.
> 
> - We don't follow why it being a plugin or not has any relevance to whether or not there is an audit trail.
> 
> - Audit trail is not a requirement for our use case.

There seems to have been some confusion about audit trail, I'm sorry if I were not clear.
Our requirements for deletion of groups are as I described previously, copy paste here:
"
1. Deleted groups go to refs/deletedgroups
2. Deleted groups have an audit trail.
3. Once a group is created with the same name as before, it finds the relevant ref in refs/deletedgroups and recreates the group, and connects the audit trails.
"

So first possibility: If those requirements work for you, we can implement it based on those requirements in core Gerrit, but that of course requires more implementation.

Second possibility: Simply take the current implementation and put it inside a (not core) plugin. We don't care about audit trail or about anything else if it's inside a (not core) plugin.

Since you don't object to it being a plugin and you also don't mind if the groups have an audit trail and can be undeleted, I have a feeling that both options are okay for you, is that right?

Patch-set: 1