Update patch set 13
Patch Set 13:

(10 comments)

> i don't really buy your comments about alternatives and cons. if you don't understand them to a point where you can write them down, then you can't claim to have actually considered them.

I have considered them with regards to refining and improving my proposal. And I have considered them with regards to not being convinced that there is a better solution overall. I think that is different from being able to write down a separate solution.

> i think you're taking a shortcut because you think you're already holding the holy grail.

It does not feel to me like I am taking a shortcut. It feels to me like going the extra mile.

I am just trying to present a solution to some use cases that I find important to address. And currently I am still convinced that this solution is the best that I can find. Whether that translates to "holding the holy grail" I don't know.

> you can't be sure that it is overall optimal

No, I can't.

> but i hope you understand that your request for comments doesn't feel quite right under such circumstances.

I don't really. I have put a lot of effort into writing this design, communicating it and iterating on it. But it is always difficult to make everyone 100% happy.
 
> i find the thought of posting a counter-proposal absurd, because it seems so utterly trivial to me - i mean, do i *really* have to explain the concept of timestamp comparisons? forum software has used this approach for decades.

Yes, you have to, because otherwise I can't comment on it and point out the shortcomings.
 
> i suppose the lack of understanding could stem from the fact that, while we both have the overarching goal to optimize the process by making the need for action clear, we have quite divergent expectations for the solution: you want participants to be passive receivers of binding action requests, while i want them to be active deciders over action opportunities.

I don't think the attention set is something passive to consume. You are absolutely expected to interact actively with the attention set.

> you may interpret the "you" as referring to google, not ben personally. doesn't change my perception of things. take it as a reminder that as an entity with a disproportionate influence google needs to work extra hard to be transparent if the open governance model is to be credible.

To me it feels that we are super transparent and we are listening carefully even to community members that don't contribute any code. What can we do better here?

Patch-set: 13