Update patch set 13
Patch Set 13: Code-Review-1

(6 comments)

> It does not feel to me like I am taking a shortcut. It feels to me like [not] going the extra mile.

surely you will agree that the distinction is merely a matter of perspective. ;)
my contention is that one really _should_ go that extra mile when things got controversial, to be sure that common understanding was reached, that it's not a "because i say so" situation, and for posterity (because it's reasonable to assume that the raised concerns were not extraordinary).

specifically, i think you should address the points pertaining to the assignee feature:
- the discussion surrounding the limitation to just one assignee. nobody challenged my conclusion, so i would suggest to adopt it.
- the unintentional straw-manning of its UI limitations as a fundamental problem. if you don't feel like considering/documenting possible solutions for them, just use something like "numerous UI implementation details limit the feature's usefulness. [optionally, you can include your list as sub-items here.] these are not of a fundamental nature, but given the other considerations, we believe that the effort is better spent elsewhere."

the other fundamental discussions are ongoing, so i won't suggest concrete edits yet.

> I don't think the attention set is something passive to consume. You are absolutely expected to interact actively with the attention set.

not in the way that i mean it, which is from the perspective of the one who will/not do the followup to the current action. with AS, others decide for you, and you decide for others in turn.

> > do i *really* have to explain the concept of timestamp comparisons?
> 
> Yes, you have to, because otherwise I can't comment on it and point out the shortcomings.

it would really help if you told me where i lost you, because as of now my proposal wouldn't be more than what i already wrote, which i don't expect to be helpful.

ok, maybe this will help:

it occurred to me, that my proposal is in fact functionally equivalent to the reviewed flag, with the  only difference being that my timestamp comparison mechanism determines the state dynamically at query time, while the flag is (de-)asserted by the previous actions of users.

it also occurred to me that your AS *is* in fact the (inverted) reviewed flag - we're merely haggling over how/when that flag is (de-)asserted and how much it influences.

my timestamp mechanism is rigid; it reliably implements an aggressive AS - but that can also be done with the flag, with some more care.

with that in mind, we can loop back to my previous conclusion that the two approaches are not mutually exclusive: we can simply have two flags, one of which is asserted conservatively (your proposal) and the other aggressively (my proposal), visualize both, and let the user configure which one they want to decide over notifications (globally and per change).

makes sense now?

Patch-set: 13
Reviewer: Gerrit User 1013267 <1013267@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1