Remove unused documentation files
Gerrit's documentation was added to the project during initial
experimentation with the new Jekyll templating system, but aside
from a few pages that are linked from the index, is not actually
published anywhere. Since it was added ithas not been kept up to
date with the current documentation that is maintained in the core
gerrit project. It's better to remove it than leave it getting
even more stale.

Change links from the index to point to the latest documentation
hosted on https://gerrit-review.googlesource.com/Documentation/

Change-Id: Id3ffac99e785750f7fae8cf6f98821dd36299e05