Make all positive votes on Code-Review label sticky
This sets copyValue=+1 and copyValue=+2 on the Code-Review label so that
all positive votes are sticky when a new patch set is uploaded.

The copyValue rule on labels is added by change Ia8d1b634b. While that
change is not submitted and rolled-out this change has no effect.

Since the definition of the Code-Review label on this repo diverges from
the default definition of the Code-Review label that is defined in
All-Projects we need to copy the full label definition into this repo.

The homepage repo does not contain source code, but documents, such as
design docs, release notes, project news etc. For these kind of
documents it is often the case that people vote with Code-Review+1 to
express their consent with the contents, but then these votes get lost
when minor issues such as nits, spelling mistakes, rephrasing, adding
some more points are addressed. Reviewers which already gave their
consent often don't come back after each iteration to re-apply their
Code-Review+1 vote, so that in the end it can look like there is not
much support for a design, while in fact many people agreed to it (and
just their Code-Review+1 votes are no longer visible). This makes it
more difficult to reason about approving a design by the ESC. To improve
this situation we want to make all positive Code-Review votes in the
homepage repo sticky.

Bug: Issue 12025
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I5bf0d7131311897c4c2883d59e78747c04334a51