Update patch set 1
Patch Set 1:

> Audit trail is required for us since it wouldn't be consistent with the rest of the REST endpoints operations if it were allowed to just delete a group without an audit trail.

OK, but that still doesn't answer why it's required when the feature is in core but not when it's in a plugin.

What about having it so that the it's optional to have the audit trail?  Or to put it another way, optional to create a backup of the deleted group.  This is similar to the delete project plugin which provides an option to move the project to another folder rather than deleting completely.

> Anyway, we will probably go with the option to make it in core with the requirements I mentioned, I hope it's okay if it takes some more time.
> 

One advantage of doing it as a plugin is that it would be possible to backport it to work with the currently released Gerrit versions rather than waiting for 3.2.

> Some notes for implementation: GroupsUpdate should have an option to delete a group. GroupsOperations should have at least 1 deletion of groups for testing purposes (both regular deletion and deletion with move to refs/deletedgroups). I will write that later in the doc.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>