Remove github-pages to reduce dependencies for generating site
The site is built with Jekyll inside Docker. On creation of the Docker
image github-pages is installed, which in turn installs Jekyll as one
of it's dependencies.

We don't actually use github-pages; installing it and all its other
dependencies adds extra time to the site generation for no reason.

Change it to only install Jekyll, which significantly reduces the
site build time.

Test plan:

  $ git clean -fdx
  $ docker-compose up
  Browse to http://localhost:4000

Change-Id: Id00da6badb79d6abfd2f6c81c21b9dbd5bd13a41