Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Seems like I don't have too much time to work on this.
> Is it going to be good for you if I migrate the code we have with its basic functionality into a plugin? Then, you will have to mostly maintain it.

I don't think we should proceed yet until we've agreed that a plugin is the way to go.  Basically we (Collabnet) don't object to it, but we still didn't get clarification on the question that I previously raised, i.e. why does having it in a plugin make any difference whether or not there is an audit trail?

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>