Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Seems like I don't have too much time to work on this.
> > Is it going to be good for you if I migrate the code we have with its basic functionality into a plugin? Then, you will have to mostly maintain it.
> 
> I don't think we should proceed yet until we've agreed that a plugin is the way to go.  Basically we (Collabnet) don't object to it, but we still didn't get clarification on the question that I previously raised, i.e. why does having it in a plugin make any difference whether or not there is an audit trail?

Let me try to answer this: Inside of core Gerrit, we support an audit trail for groups by default. There are very likely people who rely on this audit trail to work. If I was such a person, I would expect a group deletion to also play nicely together with that audit trail, especially if both are core features. Plugins, on the other hand, have the liberty to just target a specific use case, which should of course be documented. Admins can then decide whether that use case is sufficient for them or not. Applied to the group deletion feature, this means that we could have a plugin which allows group deletion without caring about the audit trail. If we ever wanted to move it to core Gerrit (or make it a core plugin), we would of course need to adjust such a plugin.

Patch-set: 1
Reviewer: Gerrit User 1030912 <1030912@173816e5-2b9a-37c3-8a2e-48639d4f1153>