Update patch set 16
Patch Set 16:

(21 comments)

> I still expect some followup on my ideas regarding the "two-level AS" and inclusion of configurability into the design.

I would appreciate, if you could leave comments on specific lines in the proposal with more concrete questions.

If you have alternative ideas that I have not incorporated as much as you would like, then please write a separate solution doc.

From https://gerrit-review.googlesource.com/Documentation/dev-design-docs.html:

"Anyone who has ideas for an alternative solution uploads a change with a solution-<n>.md that describes their solution. In case of doubt whether an idea is a refinement of an existing solution or an alternative solution, it’s up to the owner of the discussed solution to decide if the solution should be updated, or if the proposer should start a new alternative solution.

All possible solutions are fairly discussed with their pros and cons, and treated equally until a conclusion is made."

If you have ideas for improving the design doc process, then I would recommend starting a separate thread with the community, the community managers and the ESC about that.

Patch-set: 16