Update patch set 18
Patch Set 18: Code-Review+1

(2 comments)

I am in favor of the proposal to implement attention set in Gerrit.

I think Ben did an outstanding job in iterating on this doc and working with feedback. Thank you to all the reviewers who have invested their time into reading and commenting on this proposal, I think the iteration rounds were beneficial to the document and idea.

From reading the last comments, I sense that Oswald is still disagreeing with some aspects of the design. Disagreement is healthy in that it forces us to think more thoroughly about things and iterate where necessary. In an open-source project like Gerrit, that is used by many companies in many different ways, 100%-consensus is absolutely desirable, but unfortunately unrealistic to achieve at times. For the project to move forward, we have to move on at some point. From where I stand, 144 change messages and 16 iterations and what feels like consensus among the majority of contributors is a fair point in time to ask the ESC to make a decision on the proposal.

Patch-set: 18
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1