Update patch set 18
Patch Set 18: Verified+1 Code-Review+2

I am upgrading my vote to a +2, so that Ben can submit his proposed solution. This does not meant that the solution is approved or that we settled on it. This decision will be made at a later point and a 'conclusion' doc will be uploaded.

If you prefer a different solution, feel free to upload a change for review to propose it.

Patch-set: 18
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1
Label: Code-Review=+2