Update patch set 19
Patch Set 19:

(10 comments)

> I don't fully understand how the "loose" AS would work.

as implied previously, you should think of it as simply fixing the semantics of the "reviewed" flag.
or actually, a more accurate description would be "replacing the 'reviewed' flag with a 'visited' flag".
the rest are just bells and whistles, and you actually thought more about that than i had. 😄

> Would you also call users out with icons and such?

no ... well, i hadn't considered it. it's less useful than for the strict AS, but it sounds like a worthwhile refinement.

> Would you have separate dashboard sections? Would you keep bolding dashboard rows?

i'd stay with bolding. you (?) mentioned that users have problems with it, but i'd speculate that's because they perceive it as being entirely random, which is precisely the point i want to address.
anyway, i'm not committed to any particular presentation.

> How would notification emails be different?

as with the strict AS, that depends on the scope you want to set.
if we make it just about fixing "reviewed" as i originally intended, then there would be no change.
but it's also conceivable to instead send a generic "new activity on change X" (equivalently: "added to loose AS") notification once and suppress the rest until the user visits the change, as reflected in the five configuration states i suggested yesterday.
it's also possible to send occasional (batched) reminders, which is equally applicable to strict AS. that would make sense for only sporadic users of gerrit.

> Would you have a user interface for adding/removing yourself/others?

there would be definitely no separate interface to manipulate others' status, because one would do that implicitly, by virtue of "loose AS" being simply a notification about new activity by other users.

as to manipulating your own status:
- fundamentally, there should be no need to do that at all, because the flag only reflects whether you visited the change since the last modification by someone else, and the system knows that itself
- for the case where you leave the change without any observable activity, the system could pop up a query dialog whether you intend to come back, i.e., whether the flag should be really reset.
- there is already the "mark change as (not) reviewed" interface on the change screen. the "mark as reviewed" part would become obsolete, as just visiting the change would do that. i'm not sure whether an explicit "mark as not reviewed" would still make sense.
- as a strictly optional refinement, it's conceivable to add "mark as reviewed" buttons on the dashboard. to make that really useful, there would have to be tooltips which present a summary of the activity since the last visit. but note that this is also something that would make perfect sense for strict AS.

> Overall I think you would add too much complexity for too little benefit.

i disagree.
measured against the status quo, it's no additional complexity at all, because everything necessary for the minimal implementation is already in place and just needs to be re-purposed a little.
it also doesn't burden the user, because it's almost fully automatic and can be ignored by those who don't care.
measured against a strict-AS-only solution, it's obviously a bit of additional complexity, but it would mostly share facilities.

> That said, I understand where you are coming from. The problem is a bit similar to comments, which can be unresolved (strict attention). How do you make sure that resolved comments are still read (loose attention)? We don't have a great answer for that, but I would still be hesitant to add "unreplied" for loose attention, because it adds too much complexity.

i think you mean "unread".
this is indeed comparable, so let's get creative. ^^

i'll assume that it's not necessary to visualize to other users that a particular user did/not read any particular comment - that seems like a pointless UI nightmare. other than that, visualization would happen with some kind of shading like it is for resolution.
i can think of two fundamental ways to implement updating the status:
- explicitly, i.e., the user must press a "read" button below new comments which they don't handle in another way. possibly consider only threads they participate in, but i'm not sure. there is also a button "mark all a read".
- implicitly, i.e., visit-based. to make it reliable even though switches between the sub-views of a change are common, comments should be marked as read only when the change as a whole is left. huh - doesn't that sound familiar? ^^

backend-wise, i can think of two approaches again:
- add a new per-comment-per-user structure. that's effort as such, and is inherently somewhat problematic, just as the (imo useless) per-file "reviewed" flags are. note for comparison: this problem doesn't exist for two-level AS, because it's just one more bit in a structure that already exists and you want to keep anyway.
- or we come full circle: a visit-based system can be simply implemented by comparing the timestamps of the comments and the user's last visit of the change. that requires storing a timestamp instead of a single bit, but it's still the same per-change-per-user structure.

> Thanks for all your review comments!

you're welcome!
fun fact: i spent 20-30 hours on this so far. just for perspective ...

as martin, i'm disappointed that you submitted this already. it wouldn't have run away over the holidays ...

Patch-set: 19